+++ src/main/java/com/example/evostyle/domain/brand/entity/BrandCategoryLimit.java
@@ -1,4 +1,4 @@
-package com.example.evostyle.domain.brand.brandcategory;
+package com.example.evostyle.domain.brand.entity;
 
 public class BrandCategoryLimit {
     public static final int MAX_CATEGORY_COUNT = 3;
