+++ src/main/java/com/example/evostyle/domain/brand/repository/BrandCategoryRepository.java
@@ -1,7 +1,8 @@
-package com.example.evostyle.domain.brand.brandcategory;
+package com.example.evostyle.domain.brand.repository;
 
 import com.example.evostyle.domain.brand.dto.response.CategoryInfo;
 import com.example.evostyle.domain.brand.entity.Brand;
+import com.example.evostyle.domain.brand.entity.BrandCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
